Fix problem with nested processes
The problem was that when a subprocess was triggered, it had a callback to its parent's resume method. When the subprocess is triggered, it also called the resume method of the parent process, which caused the sequence to continue execution.

The solution was to introduce an event that is triggered by a process when it is finished. The difference now is that a subprocess only calls its own resume callback, which, when finishes schedules its processFinishedEvent. The processFinishedEvent has the resume callback of the parent process attached to it, so that when it is triggered by the environment, the parent process resumes **after** the subprocess is finished.